Remove `VmInterface::get_current_execution_state`
