+++ core/lib/multivm/src/versions/vm_1_3_2/vm.rs
@@ -1,31 +1,26 @@
 use std::collections::HashSet;
 
-use circuit_sequencer_api_1_3_3::sort_storage_access::sort_storage_access_queries;
-use zksync_types::{
-    l2_to_l1_log::{L2ToL1Log, UserL2ToL1Log},
-    Transaction,
-};
-use zksync_utils::{bytecode::hash_bytecode, h256_to_u256, u256_to_h256};
+use zksync_types::Transaction;
+use zksync_utils::{bytecode::hash_bytecode, h256_to_u256};
 
 use crate::{
     glue::{history_mode::HistoryMode, GlueInto},
     interface::{
         storage::{StoragePtr, WriteStorage},
-        BootloaderMemory, BytecodeCompressionError, CompressedBytecodeInfo, CurrentExecutionState,
-        FinishedL1Batch, L1BatchEnv, L2BlockEnv, SystemEnv, TxExecutionMode, VmExecutionMode,
+        BootloaderMemory, BytecodeCompressionError, CompressedBytecodeInfo, FinishedL1Batch,
+        L1BatchEnv, L2BlockEnv, SystemEnv, TxExecutionMode, VmExecutionMode,
         VmExecutionResultAndLogs, VmFactory, VmInterface, VmInterfaceHistoryEnabled,
         VmMemoryMetrics,
     },
     tracers::old::TracerDispatcher,
     utils::bytecode,
-    vm_1_3_2::{events::merge_events, VmInstance},
+    vm_1_3_2::VmInstance,
 };
 
 #[derive(Debug)]
 pub struct Vm<S: WriteStorage, H: HistoryMode> {
     pub(crate) vm: VmInstance<S, H::Vm1_3_2Mode>,
     pub(crate) system_env: SystemEnv,
-    pub(crate) batch_env: L1BatchEnv,
     pub(crate) last_tx_compressed_bytecodes: Vec<CompressedBytecodeInfo>,
 }
 
@@ -91,56 +86,6 @@ impl<S: WriteStorage, H: HistoryMode> VmInterface for Vm<S, H> {
 
     fn start_new_l2_block(&mut self, _l2_block_env: L2BlockEnv) {
         // Do nothing, because vm 1.3.2 doesn't support L2 blocks
-    }
-
-    fn get_current_execution_state(&self) -> CurrentExecutionState {
-        let (raw_events, l1_messages) = self.vm.state.event_sink.flatten();
-        let events = merge_events(raw_events)
-            .into_iter()
-            .map(|e| e.into_vm_event(self.batch_env.number))
-            .collect();
-        let l2_to_l1_logs = l1_messages
-            .into_iter()
-            .map(|m| {
-                UserL2ToL1Log(L2ToL1Log {
-                    shard_id: m.shard_id,
-                    is_service: m.is_first,
-                    tx_number_in_block: m.tx_number_in_block,
-                    sender: m.address,
-                    key: u256_to_h256(m.key),
-                    value: u256_to_h256(m.value),
-                })
-            })
-            .collect();
-
-        let used_contract_hashes = self
-            .vm
-            .state
-            .decommittment_processor
-            .known_bytecodes
-            .inner()
-            .keys()
-            .cloned()
-            .collect();
-
-        let storage_log_queries = self.vm.state.storage.get_final_log_queries();
-
-        let deduped_storage_log_queries =
-            sort_storage_access_queries(storage_log_queries.iter().map(|log| &log.log_query)).1;
-
-        CurrentExecutionState {
-            events,
-            deduplicated_storage_logs: deduped_storage_log_queries
-                .into_iter()
-                .map(GlueInto::glue_into)
-                .collect(),
-            used_contract_hashes,
-            user_l2_to_l1_logs: l2_to_l1_logs,
-            system_logs: vec![],
-            // Fields below are not produced by VM 1.3.2
-            storage_refunds: vec![],
-            pubdata_costs: Vec::new(),
-        }
     }
 
     fn inspect_transaction_with_bytecode_compression(
@@ -270,7 +215,7 @@ impl<S: WriteStorage, H: HistoryMode> VmFactory<S> for Vm<S, H> {
         let inner_vm: VmInstance<S, H::Vm1_3_2Mode> =
             crate::vm_1_3_2::vm_with_bootloader::init_vm_with_gas_limit(
                 oracle_tools,
-                batch_env.clone().glue_into(),
+                batch_env.glue_into(),
                 block_properties,
                 system_env.execution_mode.glue_into(),
                 &system_env.base_system_smart_contracts.clone().glue_into(),
@@ -279,7 +224,6 @@ impl<S: WriteStorage, H: HistoryMode> VmFactory<S> for Vm<S, H> {
         Self {
             vm: inner_vm,
             system_env,
-            batch_env,
             last_tx_compressed_bytecodes: vec![],
         }
     }
