+++ core/lib/multivm/src/versions/vm_1_4_2/vm.rs
@@ -38,40 +38,7 @@ pub struct Vm<S: WriteStorage, H: HistoryMode> {
     _phantom: std::marker::PhantomData<H>,
 }
 
-impl<S: WriteStorage, H: HistoryMode> VmInterface for Vm<S, H> {
-    type TracerDispatcher = TracerDispatcher<S, H::Vm1_4_2>;
-
-    /// Push tx into memory for the future execution
-    fn push_transaction(&mut self, tx: Transaction) {
-        self.push_transaction_with_compression(tx, true);
-    }
-
-    /// Execute VM with custom tracers.
-    fn inspect(
-        &mut self,
-        tracer: Self::TracerDispatcher,
-        execution_mode: VmExecutionMode,
-    ) -> VmExecutionResultAndLogs {
-        self.inspect_inner(tracer, execution_mode, None)
-    }
-
-    /// Get current state of bootloader memory.
-    fn get_bootloader_memory(&self) -> BootloaderMemory {
-        self.bootloader_state.bootloader_memory()
-    }
-
-    /// Get compressed bytecodes of the last executed transaction
-    fn get_last_tx_compressed_bytecodes(&self) -> Vec<CompressedBytecodeInfo> {
-        self.bootloader_state.get_last_tx_compressed_bytecodes()
-    }
-
-    fn start_new_l2_block(&mut self, l2_block_env: L2BlockEnv) {
-        self.bootloader_state.start_new_l2_block(l2_block_env);
-    }
-
-    /// Get current state of virtual machine.
-    /// This method should be used only after the batch execution.
-    /// Otherwise it can panic.
+impl<S: WriteStorage, H: HistoryMode> Vm<S, H> {
     fn get_current_execution_state(&self) -> CurrentExecutionState {
         let (raw_events, l1_messages) = self.state.event_sink.flatten();
         let events: Vec<_> = merge_events(raw_events)
@@ -106,8 +73,38 @@ impl<S: WriteStorage, H: HistoryMode> VmInterface for Vm<S, H> {
             pubdata_costs: Vec::new(),
         }
     }
+}
 
-    /// Execute transaction with optional bytecode compression.
+impl<S: WriteStorage, H: HistoryMode> VmInterface for Vm<S, H> {
+    type TracerDispatcher = TracerDispatcher<S, H::Vm1_4_2>;
+
+    /// Push tx into memory for the future execution
+    fn push_transaction(&mut self, tx: Transaction) {
+        self.push_transaction_with_compression(tx, true);
+    }
+
+    /// Execute VM with custom tracers.
+    fn inspect(
+        &mut self,
+        tracer: Self::TracerDispatcher,
+        execution_mode: VmExecutionMode,
+    ) -> VmExecutionResultAndLogs {
+        self.inspect_inner(tracer, execution_mode, None)
+    }
+
+    /// Get current state of bootloader memory.
+    fn get_bootloader_memory(&self) -> BootloaderMemory {
+        self.bootloader_state.bootloader_memory()
+    }
+
+    /// Get compressed bytecodes of the last executed transaction
+    fn get_last_tx_compressed_bytecodes(&self) -> Vec<CompressedBytecodeInfo> {
+        self.bootloader_state.get_last_tx_compressed_bytecodes()
+    }
+
+    fn start_new_l2_block(&mut self, l2_block_env: L2BlockEnv) {
+        self.bootloader_state.start_new_l2_block(l2_block_env);
+    }
 
     /// Inspect transaction with optional bytecode compression.
     fn inspect_transaction_with_bytecode_compression(
